Update array access for compatibility with numpy 1.12
Recent numpy versions deprecated using anything apart from a int to index a numpy
array. This change explictly sets the correct data type for the array indexes.

+++ supremm/subsample.py
@@ -13,7 +13,7 @@ class TimeseriesAccumulator(object):
         self._samplewindow = None
         self._leadout = None
         self._data = numpy.empty((nhosts, TimeseriesAccumulator.MAX_DATAPOINTS, 2))
-        self._count = numpy.zeros(nhosts)
+        self._count = numpy.zeros(nhosts, dtype=int)
 
     def adddata(self, hostidx, timestamp, value):
         """ Add a datapoint to the collection.
